Refactor and clean.
