+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp/src/test/java/org/kie/workbench/common/dmn/showcase/client/screens/editor/SessionDiagramEditorScreenTest.java
@@ -24,7 +24,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.kie.workbench.common.dmn.client.decision.DecisionNavigatorDock;
 import org.kie.workbench.common.dmn.client.editors.expressions.ExpressionEditorView;
-import org.kie.workbench.common.dmn.client.graph.DMNGraphLayout;
 import org.kie.workbench.common.dmn.client.session.DMNEditorSession;
 import org.kie.workbench.common.dmn.showcase.client.perspectives.AuthoringPerspective;
 import org.kie.workbench.common.stunner.client.widgets.presenters.session.SessionPresenter;
@@ -35,8 +34,8 @@ import org.kie.workbench.common.stunner.core.client.canvas.AbstractCanvasHandler
 import org.kie.workbench.common.stunner.core.client.session.impl.EditorSession;
 import org.kie.workbench.common.stunner.core.diagram.Diagram;
 import org.kie.workbench.common.stunner.core.diagram.Metadata;
+import org.kie.workbench.common.stunner.core.graph.processing.layout.LayoutService;
 import org.kie.workbench.common.stunner.forms.client.event.RefreshFormPropertiesEvent;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.AutomaticLayoutService;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
@@ -83,10 +82,7 @@ public class SessionDiagramEditorScreenTest {
     private ArgumentCaptor<Consumer<EditorSession>> clientFullSessionConsumer;
 
     @Mock
-    private DMNGraphLayout dmnGraphLayout;
-
-    @Mock
-    private AutomaticLayoutService layoutService;
+    private LayoutService layoutService;
 
     private SessionDiagramEditorScreen editor;
 
@@ -126,8 +122,7 @@ public class SessionDiagramEditorScreenTest {
                                                     screenPanelView,
                                                     null,
                                                     decisionNavigatorDock,
-                                                    layoutService,
-                                                    dmnGraphLayout));
+                                                    layoutService));
     }
 
     @Test
