--- kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-api/kie-wb-common-stunner-core-api/src/main/java/org/kie/workbench/common/stunner/core/graph/processing/layout/step02/LongestPathVertexLayerer.java
@@ -1,102 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.workbench.common.stunner.core.graph.processing.layout.step02;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.NoSuchElementException;
-
-import javax.enterprise.inject.Default;
-
-import org.kie.workbench.common.stunner.core.graph.processing.layout.Layer;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.ReorderedGraph;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.Vertex;
-
-/**
- * Assign each vertex in a graph to a layer, using the longest path algorithm.
- */
-@Default
-public final class LongestPathVertexLayerer implements VertexLayerer {
-
-    private Vertex[] vertices;
-    private final HashMap<String, Integer> vertexHeight;
-    private ReorderedGraph graph;
-
-    public LongestPathVertexLayerer() {
-        this.vertexHeight = new HashMap<>();
-    }
-
-    /**
-     * Create layers for the graph and assign each vertex to a layer.
-     * @param graph The graph.
-     */
-    @Override
-    public void createLayers(final ReorderedGraph graph) {
-        this.graph = graph;
-        this.vertices = new Vertex[graph.getVertices().size()];
-
-        for (int i = 0; i < graph.getVertices().size(); i++) {
-            final String v = graph.getVertices().get(i);
-            this.vertices[i] = new Vertex(v);
-            this.vertexHeight.put(v, -1);
-        }
-
-        for (final Vertex vertex : this.vertices) {
-            visit(vertex);
-        }
-    }
-
-    private int visit(final Vertex vertex) {
-        final int height = this.vertexHeight.getOrDefault(vertex.getId(), 0);
-        if (height >= 0) {
-            return height;
-        }
-
-        int maxHeight = 1;
-
-        final String[] verticesFromHere = graph.getVerticesFrom(vertex.getId());
-        for (final String nextVertex : verticesFromHere) {
-            if (!nextVertex.equals(vertex.getId())) {
-                final Vertex next = Arrays.stream(this.vertices)
-                        .filter(f -> f.getId().equals(nextVertex))
-                        .findFirst()
-                        .orElseThrow(() -> new NoSuchElementException(
-                                "Can not found the vertex pointed in other side of the Edge."
-                        ));
-
-                final int targetHeight = visit(next);
-                maxHeight = Math.max(maxHeight, targetHeight + 1);
-            }
-        }
-
-        addToLayer(vertex, maxHeight);
-        return maxHeight;
-    }
-
-    private void addToLayer(final Vertex vertex,
-                            final int height) {
-        for (int i = this.graph.getLayers().size(); i < height; i++) {
-            this.graph.getLayers().add(0, new Layer());
-        }
-
-        final int level = this.graph.getLayers().size() - height;
-        final Layer layer = this.graph.getLayers().get(level);
-        layer.setLevel(height);
-        layer.addVertex(vertex);
-        vertexHeight.put(vertex.getId(), height);
-    }
-}
