+++ kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-api/kie-wb-common-stunner-core-api/src/main/java/org/kie/workbench/common/stunner/core/graph/processing/layout/sugiyama/step02/VertexLayerer.java
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package org.kie.workbench.common.stunner.core.graph.processing.layout;
+package org.kie.workbench.common.stunner.core.graph.processing.layout.sugiyama.step02;
 
-import java.util.List;
+import org.kie.workbench.common.stunner.core.graph.processing.layout.ReorderedGraph;
 
-import org.kie.workbench.common.stunner.core.graph.processing.layout.sugiyama.OrientedEdge;
+public interface VertexLayerer {
 
-public interface ReorderedGraph {
-
-    List<OrientedEdge> getEdges();
+    void createLayers(ReorderedGraph graph);
 }
