--- kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-api/kie-wb-common-stunner-core-api/src/test/java/org/kie/workbench/common/stunner/core/graph/processing/layout/step03/VertexOrderingTest.java
@@ -1,405 +1,0 @@
-/*
- * Copyright 2018 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.workbench.common.stunner.core.graph.processing.layout.step03;
-
-import java.util.ArrayList;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.Edge;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.Layer;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.ReorderedGraph;
-import org.kie.workbench.common.stunner.core.graph.processing.layout.Vertex;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-
-@RunWith(MockitoJUnitRunner.class)
-public class VertexOrderingTest {
-
-    @Test
-    public void testSimpleReorder() {
-        final ReorderedGraph graph = new ReorderedGraph();
-        graph.addEdge("A", "D");
-        graph.addEdge("B", "C");
-
-        final Layer layer01 = new Layer(1);
-        layer01.addNewVertex("A");
-        layer01.addNewVertex("B");
-        graph.getLayers().add(layer01);
-
-        final Layer layer02 = new Layer(2);
-        layer02.addNewVertex("C");
-        layer02.addNewVertex("D");
-        graph.getLayers().add(layer02);
-
-        final MedianVertexLayerPositioning median = new MedianVertexLayerPositioning();
-        final LayerCrossingCount layersCount = new LayerCrossingCount();
-        final VerticesTransposer verticesTransposer = new VerticesTransposer(layersCount);
-        final DefaultVertexOrdering ordering = new DefaultVertexOrdering(median,
-                                                                         layersCount,
-                                                                         verticesTransposer);
-        ordering.orderVertices(graph);
-
-        assertArrayEquals(new Object[]{"A", "B"}, graph.getLayers().get(0).getVertices().stream().map(Vertex::getId).toArray());
-        assertArrayEquals(new Object[]{"D", "C"}, graph.getLayers().get(1).getVertices().stream().map(Vertex::getId).toArray());
-    }
-
-    @Test
-    public void testReorder() {
-        final ReorderedGraph graph = new ReorderedGraph();
-        graph.addEdge("A", "B");
-        graph.addEdge("A", "C");
-        graph.addEdge("F", "B");
-        graph.addEdge("C", "E");
-        graph.addEdge("G", "C");
-        graph.addEdge("C", "H");
-        graph.addEdge("D", "F");
-
-        final Layer layer01 = new Layer(1);
-        layer01.addNewVertex("A");
-        graph.getLayers().add(layer01);
-
-        final Layer layer02 = new Layer(2);
-        layer02.addNewVertex("B");
-        layer02.addNewVertex("C");
-        layer02.addNewVertex("D");
-        graph.getLayers().add(layer02);
-
-        final Layer layer03 = new Layer(3);
-        layer03.addNewVertex("E");
-        layer03.addNewVertex("F");
-        layer03.addNewVertex("G");
-        layer03.addNewVertex("H");
-        graph.getLayers().add(layer03);
-
-        final MedianVertexLayerPositioning median = new MedianVertexLayerPositioning();
-        final LayerCrossingCount layersCount = new LayerCrossingCount();
-        final VerticesTransposer verticesTransposer = new VerticesTransposer(layersCount);
-        final DefaultVertexOrdering ordering = new DefaultVertexOrdering(median,
-                                                                         layersCount,
-                                                                         verticesTransposer);
-        ordering.orderVertices(graph);
-        final ArrayList<Layer> orderedLayers = graph.getLayers();
-
-        assertArrayEquals(new Object[]{"A"}, orderedLayers.get(0).getVertices().stream().map(Vertex::getId).toArray());
-        assertArrayEquals(new Object[]{"D", "B", "C"}, orderedLayers.get(1).getVertices().stream().map(Vertex::getId).toArray());
-        assertArrayEquals(new Object[]{"F", "E", "G", "H"}, orderedLayers.get(2).getVertices().stream().map(Vertex::getId).toArray());
-    }
-
-    @Test
-    public void calculateMedianTest() {
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("G", "A"));
-        edges.add(new Edge("G", "D"));
-        edges.add(new Edge("G", "E"));
-
-        final Layer layer00 = new Layer(0);
-        layer00.addNewVertex("A");
-        layer00.addNewVertex("B");
-        layer00.addNewVertex("C");
-        layer00.addNewVertex("D");
-        layer00.addNewVertex("E");
-        final Layer layer01 = new Layer(1);
-        layer01.addNewVertex("F");
-        layer01.addNewVertex("G");
-        layer01.addNewVertex("H");
-        layer01.addNewVertex("I");
-        layer01.addNewVertex("J");
-
-        final MedianVertexLayerPositioning vertexMedian = new MedianVertexLayerPositioning();
-        final double median = vertexMedian.calculateMedianOfVerticesConnectedTo("G", layer00, edges);
-        assertEquals(3.0, median, 0.0001);
-    }
-
-    @Test
-    public void testSimpleCrossing() {
-
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("C");
-        bottom.addNewVertex("D");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "D"));
-        edges.add(new Edge("B", "C"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(1, result);
-    }
-
-    @Test
-    public void testSimpleNoCrossing() {
-
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("C");
-        bottom.addNewVertex("D");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "C"));
-        edges.add(new Edge("B", "D"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(0, result);
-    }
-
-    @Test
-    public void test1Crossing() {
-
-        /*
-         * 1 crossing
-         * A   B   C
-         *   \   /
-         *     X
-         *   /   \
-         * D   E   F
-         */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-        top.addNewVertex("C");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "F"));
-        edges.add(new Edge("D", "C"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(1, result);
-    }
-
-    @Test
-    public void test2CrossingsUnevenLayers() {
-
-        /*
-         * 2 crossings
-         *       A           B
-         *      /\\         /
-         *     /  \ -------+ --
-         *    / /-- +-----/    \
-         *   / /     \          \
-         *  D         E          F
-         * */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "D"));
-        edges.add(new Edge("E", "A"));
-        edges.add(new Edge("A", "F"));
-        edges.add(new Edge("D", "B"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(2, result);
-    }
-
-    @Test
-    public void test2Crossings() {
-
-        /*
-         * 2 crossings
-         * A   B   C
-         *  \  | /
-         *   x x
-         *  / \|
-         * D   E   F
-         * */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-        top.addNewVertex("C");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "E"));
-        edges.add(new Edge("B", "E"));
-        edges.add(new Edge("C", "D"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(2, result);
-    }
-
-    @Test
-    public void test2Crossing8Vertex() {
-
-        /*
-         * 2 crossing
-         *  A       B       C        D
-         *  |\--\   |        \---\   |
-         *  | \  \--+-------\     \  |
-         *  |  \----+-----\  \     \ |
-         *  |       |      \  \---\ \|
-         *  E       F       G        H
-         */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-        top.addNewVertex("C");
-        top.addNewVertex("D");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-        bottom.addNewVertex("G");
-        bottom.addNewVertex("H");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "E"));
-        edges.add(new Edge("A", "G"));
-        edges.add(new Edge("A", "H"));
-        edges.add(new Edge("B", "F"));
-        edges.add(new Edge("C", "H"));
-        edges.add(new Edge("D", "H"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(2, result);
-    }
-
-    @Test
-    public void test3CrossingsInMiddle() {
-
-        /*
-         * 3 crossings
-         * A   B   C
-         *   \ |  /
-         *     X
-         *   / | \
-         * D   E   F
-         */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-        top.addNewVertex("C");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "F"));
-        edges.add(new Edge("B", "E"));
-        edges.add(new Edge("C", "D"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(3, result);
-    }
-
-    @Test
-    public void testK33GraphCrossing() {
-
-        /*
-         * k33 - every vertex from layer 1 connected to every vertex in layer 2
-         */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-        top.addNewVertex("C");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "D"));
-        edges.add(new Edge("A", "E"));
-        edges.add(new Edge("A", "F"));
-
-        edges.add(new Edge("B", "D"));
-        edges.add(new Edge("B", "E"));
-        edges.add(new Edge("B", "F"));
-
-        edges.add(new Edge("C", "D"));
-        edges.add(new Edge("C", "E"));
-        edges.add(new Edge("C", "F"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(9, result);
-    }
-
-    @Test
-    public void test5VertexUnevenLayersNoCrossing() {
-
-        /*
-         * 3 crossings
-         * A    B
-         * |   /|\
-         * |  / | \
-         * | /  |  \
-         * D    E   F
-         */
-        final Layer top = new Layer(0);
-        top.addNewVertex("A");
-        top.addNewVertex("B");
-
-        final Layer bottom = new Layer(1);
-        bottom.addNewVertex("D");
-        bottom.addNewVertex("E");
-        bottom.addNewVertex("F");
-
-        final ArrayList<Edge> edges = new ArrayList<>();
-        edges.add(new Edge("A", "D"));
-        edges.add(new Edge("B", "D"));
-        edges.add(new Edge("E", "B"));
-        edges.add(new Edge("F", "B"));
-
-        final LayerCrossingCount cc = new LayerCrossingCount();
-        final int result = cc.crossing(edges, top, bottom);
-
-        assertEquals(0, result);
-    }
-}
